adding ensemble builder
