--- python_src/models_RF_GBM_B/RF_GBM_resultsB.csv
@@ -1,10 +1,0 @@
-,RF_precision,RF_recall,RF_thresh,GBM_precision,GBM_recall,GBM_thresh
-2006,0.45454545454545453,0.2925531914893617,0.4800056027816513,0.46794871794871795,0.3882978723404255,5.85808451969324
-2007,0.4506172839506173,0.46496815286624205,0.48538486206879194,0.45185185185185184,0.3885350318471338,5.707268111539694
-2008,0.2608695652173913,0.125,0.514571264384297,0.45614035087719296,0.18055555555555555,6.04923378039547
-2009,0.4576271186440678,0.25471698113207547,0.5223766881713169,0.46616541353383456,0.29245283018867924,5.814456357687864
-2010,0.4358974358974359,0.0913978494623656,0.6442629535332243,0.375,0.06451612903225806,6.162203676602277
-2011,0.4,0.07228915662650602,0.6028022473809834,0.46153846153846156,0.18072289156626506,6.091510041452577
-2012,0.2641509433962264,0.07329842931937172,0.5717644511192105,0.36363636363636365,0.06282722513089005,6.2730485099018765
-2013,0.45263157894736844,0.22631578947368422,0.5019779491628155,0.4827586206896552,0.07368421052631578,6.219730080163715
-2014,0.45652173913043476,0.1044776119402985,0.5441756719186722,0.4583333333333333,0.16417910447761194,5.990670837457025
